pyproject.toml
[project]
name = "postgres"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "python-dotenv>=1.0.0",
    "fastapi>=0.131.0",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.46",
    "sqlglot>=28.10.1",
    "uvicorn>=0.41.0",
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
